redirect to Feed after Filter Activate

+++ player/views.py
@@ -254,21 +254,26 @@ class FiltersView(FeedView):
             context |= super().context(request, *args, **kwargs)
             return render(request, self.template_name, context)
         logger.warn(context)
-        activate = context.get('ACTIVATE', [])
-        logger.info(f'activating filters {activate}')
-        for filter in Filter.objects.filter(player=player):
-            try:
-                filter.active = str(filter.id) in activate
-                filter.save()
-            except:
-                logger.exception('failed to activate filter: {} : {}'.format(player, filter.id))
-        for filter_code in context.get('DELETE', []):
-            try:
-                junk = Filter.objects.get(pk=filter_code)
-                logger.info(f'Deleting filter: {junk}')
-                junk.delete()
-            except:
-                logger.exception('failed to delete filter: {} : {}'.format(player, filter_code))
+        if 'ACTIVATE' in context:
+            activate = context.get('ACTIVATE', [])
+            logger.info(f'activating filters {activate}')
+            for filter in Filter.objects.filter(player=player):
+                try:
+                    filter.active = str(filter.id) in activate
+                    filter.save()
+                except:
+                    logger.exception('failed to activate filter: {} : {}'.format(player, filter.id))
+            return HttpResponseRedirect("/player/feed/")
+        if 'DELETE' in context:
+            delete = context.get('DELETE', [])
+            logger.info(f'deleting filters {delete}')
+            for filter_code in delete:
+                try:
+                    junk = Filter.objects.get(pk=filter_code)
+                    logger.info(f'Deleting filter: {junk}')
+                    junk.delete()
+                except:
+                    logger.exception('failed to delete filter: {} : {}'.format(player, filter_code))
         return HttpResponseRedirect("/player/feed/filters")
